refactor: move ValidatedCluster structs to controller/mod.rs
